requirements.txt
aiohttp
appdirs
async-timeout
atomicwrites
attrs
Automat
cachetools
certifi
cffi
chardet
click
colorama
constantly
cryptography
cssselect
cycler
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-renderer
dash-table
distlib
filelock
Flask
Flask-Compress
future
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-core
google-cloud-datastore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
httplib2
hyperlink
idna
importlib-metadata
incremental
itemadapter
itemloaders
itsdangerous
Jinja2
jmespath
lxml
MarkupSafe
more-itertools
multidict
numpy
oauth2client
oauthlib
packaging
pandas
pandas-datareader
parsel
plotly
pluggy
Protego
protobuf
py
pyasn1
pyasn1-modules
pycparser
PyDispatcher
PyHamcrest
pyOpenSSL
pyparsing
pytest
python-dateutil
pytz
queuelib
requests
requests-oauthlib
retrying
rsa
Scrapy
service-identity
six
Twisted
typing-extensions
uritemplate
urllib3
virtualenv
w3lib
wcwidth
Werkzeug
yarl
zipp
zope.interface
